Fixed serialization issue with Charity.last_year_donation_total.cents

+++ everydayhero.Api/v3/Entities/LastYearDonationTotal.cs
@@ -2,7 +2,7 @@ namespace everydayhero.Api.v3
 {
     public class LastYearDonationTotal
     {
-        public int cents { get; set; }
+        public long cents { get; set; }
         public Currency currency { get; set; }
     }
 }
